backend/requirements.txt
fastapi
uvicorn
anthropic
chromadb
langchain
python-dotenv
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
